Create Swap_adjancent_in_LR_string.cpp
In a string composed of 'L', 'R', and 'X' characters, like

"RXXLRXRXL", a move consists of either replacing one

occurrence of "XL" with "LX", or replacing one occurrence

of "RX" with "XR". Given the starting string start and the

ending string result, return True if and only if there exists a

sequence of moves to transform start to result.