Create Bag_of_tokens.cpp
You start with an initial power of power, an initial score of o, and a bag of tokens given as an integer array tokens, where each tokens [il denotes the value of token;.
Your goal is to maximize the total score by strategically playing these tokens. In one move, you can play an unplayed token in one of the two ways (but not both for the same token):
• Face-up: If your current power is at least tokens [11, you may play token;, losing tokens [il power and gaining 1 score.
• Face-down: If your current score is at least 1, you may play token;, gaining tokens [i] power and losing 1 score.
Return the maximum possible score you can achieve after playing any number of tokens.